Task/frontend refactoring

+++ backend/robotarm/src/main/java/com/sundtrack/robotarm/sequence/model/Pose.java
@@ -5,35 +5,35 @@ import jakarta.persistence.Embeddable;
 @Embeddable
 public class Pose {
 
-    private double x;
-    private double y;
-    private double z;
-    private double roll;
-    private double pitch;
-    private double yaw;
+    private long x;
+    private long y;
+    private long z;
+    private long roll;
+    private long pitch;
+    private long yaw;
 
     // Getters and Setters
-    public double getX() { return x; }
+    public long getX() { return x; }
 
-    public void setX(double x) { this.x = x; }
+    public void setX(long x) { this.x = x; }
 
-    public double getY() { return y; }
+    public long getY() { return y; }
 
-    public void setY(double y) { this.y = y; }
+    public void setY(long y) { this.y = y; }
 
-    public double getZ() { return z; }
+    public long getZ() { return z; }
 
-    public void setZ(double z) { this.z = z; }
+    public void setZ(long z) { this.z = z; }
 
-    public double getRoll() { return roll; }
+    public long getRoll() { return roll; }
 
-    public void setRoll(double roll) { this.roll = roll; }
+    public void setRoll(long roll) { this.roll = roll; }
 
-    public double getPitch() { return pitch; }
+    public long getPitch() { return pitch; }
 
-    public void setPitch(double pitch) { this.pitch = pitch; }
+    public void setPitch(long pitch) { this.pitch = pitch; }
 
-    public double getYaw() { return yaw; }
+    public long getYaw() { return yaw; }
 
-    public void setYaw(double yaw) { this.yaw = yaw; }
+    public void setYaw(long yaw) { this.yaw = yaw; }
 }
